front layout, update controllers and services

+++ src/main/java/org/example/evangelhistory/services/MediaAlbumService.java
@@ -5,38 +5,58 @@ import org.example.evangelhistory.entities.MediaAlbum;
 import org.example.evangelhistory.entities.Photo;
 import org.example.evangelhistory.repositories.MediaAlbumRepository;
 import org.example.evangelhistory.repositories.PhotoRepository;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class MediaAlbumService {
 
     private final MediaAlbumRepository albumRepo;
-    private final PhotoRepository photoRepo;
 
     public List<MediaAlbum> getAllAlbums() {
         return albumRepo.findAll();
     }
 
     public MediaAlbum getAlbum(Long id) {
-        return albumRepo.findById(id).orElseThrow(() -> new RuntimeException("Альбом не найден"));
+        Optional<MediaAlbum> albumFromDB = albumRepo.findById(id);
+        if (albumFromDB.isPresent()) {
+           return albumFromDB.get();
+        } else {
+            throw new NullPointerException("Альбом не найден");
+        }
+    }
+    public MediaAlbum findByTitle(String title) {
+        MediaAlbum albumFromDB = albumRepo.findMediaAlbumByTitle(title);
+        if (albumFromDB != null) {
+            return albumFromDB;
+        } else {
+            throw new NullPointerException("Альбом не найден");
+        }
+    }
+    public void save (MediaAlbum album) {
+        albumRepo.save(album);
     }
 
-    public MediaAlbum createAlbum(MediaAlbum album) {
-        album.setCreatedAt(java.time.LocalDateTime.now());
-        return albumRepo.save(album);
+    public void addPhoto (@NotNull Long id, Photo photo){
+        Optional<MediaAlbum> albumFromDB = albumRepo.findById(id);
+        if (albumFromDB.isPresent()) {
+            albumFromDB.get().getPhotos().add(photo);
+        } else {
+            throw new NullPointerException("Альбом не найден");
+        }
     }
 
-    public MediaAlbum addImageToAlbum (Long albumId, Photo photo){
-        MediaAlbum album = getAlbum(albumId);
-        album.addPhoto(photo);
-        return albumRepo.save(album);
-    }
-
-    public void deleteAlbum(Long id) {
-        albumRepo.deleteById(id);
+    public void deleteById(@NotNull Long id) {
+        Optional<MediaAlbum> albumFromDB = albumRepo.findById(id);
+        if (albumFromDB.isPresent()) {
+            albumRepo.deleteById(id);
+        } else {
+            throw new NullPointerException("Альбом не найден");
+        }
     }
 
 }
